+++ packages/webview_flutter/webview_flutter_platform_interface/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 2.13.0
+
+* Adds support to respond to recoverable SSL certificate errors. See `PlatformNavigationDelegate.setOnSSlAuthError`.
+
+## 2.12.0
+
+* Adds support to set whether to draw the scrollbar. See
+  `PlatformWebViewController.setVerticalScrollBarEnabled`,
+  `PlatformWebViewController.setHorizontalScrollBarEnabled`,
+  `PlatformWebViewController.supportsSetScrollBarsEnabled`.
+
 ## 2.11.0
 
 * Adds support to set the over-scroll mode for the WebView. See `PlatformWebViewController.setOverScrollMode`.